Removes index parsing from header.
Initially I was going to parse the index/barcode from the FASTQ header
strings and make sure they were all the same. However, they are rarely
the same due to mismatches and 'N'. I decided to remove this feature
because it was time consuming, confusing, and mad the output table
large.